move back-button handling to base class ContactSelectionActivity
