+++ src/org/thoughtcrime/securesms/NewConversationActivity.java
@@ -22,7 +22,6 @@ import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
@@ -184,17 +183,6 @@ public class NewConversationActivity extends ContactSelectionActivity {
     }
     startActivity(intent);
     finish();
-  }
-
-  @Override
-  public boolean onOptionsItemSelected(MenuItem item) {
-    super.onOptionsItemSelected(item);
-
-    switch (item.getItemId()) {
-    case android.R.id.home:   super.onBackPressed(); return true;
-    }
-
-    return false;
   }
 
   @Override
